Emacs: detect the phisical size of the screen more accurately
When run on Linux with multiple displays, Emacs reports the
combined width of the displays as size. This new method picks
up the details from the display the frame is located on.

Also adjust the fonts to pleasant sizes on all combinations
of internal and large external screens on both Mac and Linux.